Remove android-ktx and move it to androidMain source set
