--- android-ktx/README.md
@@ -1,14 +1,0 @@
-# Trikot.kword Android Kotlin Extensions
-
-### Installation
-Add dependency
-```groovy
-dependencies {
-    implementation "com.mirego.trikot.kword:android-ktx:$trikot_kword_android_ktx_version"
-}
-```
-
-### Change current language
-```kotlin
-    AndroidKWord.setCurrentLanguage("en")
-```
