--- android-ktx/build.gradle
@@ -1,59 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'kotlin-platform-android'
-    id 'digital.wup.android-maven-publish' version '3.6.2'
-    id 'mirego.release' version '2.0'
-    id 'mirego.publish' version '1.0'
-}
-
-repositories {
-    google()
-    jcenter()
-    mavenLocal()
-    mavenCentral()
-    maven { url "https://kotlin.bintray.com/kotlinx" }
-    maven { url 'https://jitpack.io' }
-    maven { url('https://s3.amazonaws.com/mirego-maven/public') }
-}
-
-group 'com.mirego.trikot.kword'
-
-android {
-    compileSdkVersion 30
-
-    defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 30
-    }
-}
-
-configurations.all {
-    exclude group: "org.reactivestreams"
-}
-
-dependencies {
-    implementation 'com.mirego.trikot:kword:1.0.5'
-    implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:$kotlinx_serialization_version"
-}
-
-release {
-    checkTasks = ['check']
-    buildTasks = ['publish']
-    updateVersionPart = 2
-    tagPrefix = 'android-ktx-'
-}
-
-task sourcesJar(type: Jar) {
-    from android.sourceSets.main.java.srcDirs
-    archiveClassifier.set('sources')
-}
-
-publishing {
-    publications {
-        mavenAar(MavenPublication) {
-            from components.android
-            artifact sourcesJar
-        }
-    }
-}
