--- android-ktx/gradle.properties
@@ -1,2 +1,0 @@
-#Wed Feb 03 12:42:44 EST 2021
-version=1.0.3-SNAPSHOT
