+++ gradle.properties
@@ -1,6 +1,7 @@
 kotlin.code.style=official
 kotlin_version=1.4.30
 kotlinx_serialization_version=1.0.1
+kotlin.mpp.stability.nowarn=true
 kotlin.mpp.enableGranularSourceSetsMetadata=true
 kotlin.native.enableDependencyPropagation=false
 trikot_foundation_version=2.0.0
