+++ jenkins-jobs/plugin_release.groovy
@@ -83,42 +83,3 @@ job(context.jobFullName) {
         }
     }
 }
-
-job("$context.jobFullName-android-ktx") {
-    description("Make a release of trikot.kword.android-ktx")
-    logRotator(5)
-    scm {
-        git {
-            branch("${GIT_BRANCH}")
-            remote {
-                name('origin')
-                url("${GIT_URL}")
-                credentials('github')
-            }
-        }
-    }
-    wrappers {
-        credentialsBinding {
-            amazonWebServicesCredentialsBinding {
-                accessKeyVariable('MAVEN_AWS_KEY')
-                secretKeyVariable('MAVEN_AWS_SECRET')
-                credentialsId('mirego-maven-aws')
-            }
-        }
-    }
-    steps {
-        gradle {
-            useWrapper()
-            makeExecutable()
-            tasks(':android-ktx:release')
-            switches('-i -s')
-        }
-    }
-    publishers {
-        slackNotifier {
-            notifyFailure(true)
-            notifyBackToNormal(true)
-            room(context.slackChannel)
-        }
-    }
-}
